fix : avoid filtering configuration files EXO-59810
After adding the term -exo- to the version of the eXo modules, Gatein-portal and commons unit tests were failing. The problem cause was an errouneous filtering in the kernel module that modify the configuration.xml path when its path contains **exo.** or **exo-** , this leads to creating a PortalContainer with missing services configurations. This code was removed as it is not viable and was added just to workaround an old problem with Jboss AS servers which are no more supported since a while